backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
Pillow==10.2.0
google-generativeai==0.3.2
pydantic==2.6.1
python-multipart==0.0.9